+++ tests/app/test_dependencies.py
@@ -4,34 +4,47 @@ import json
 import os
 from pathlib import Path
 from typing import AsyncIterator
-from unittest.mock import Mock
+from unittest.mock import Mock, patch
 
 import pytest
+from fastapi import Request
+from fastapi.exceptions import HTTPException
 from httpx import AsyncClient
 from langchain_openai import ChatOpenAI
 from langgraph.checkpoint.postgres.aio import AsyncPostgresSaver
 from langgraph.checkpoint.sqlite.aio import AsyncSqliteSaver
+from sqlalchemy import create_engine
+from sqlalchemy.orm import Session
 
 from neuroagent.agents import SimpleAgent, SimpleChatAgent
 from neuroagent.app.dependencies import (
     Settings,
     get_agent,
     get_agent_memory,
+    get_bluenaas_tool,
     get_brain_region_resolver_tool,
     get_cell_types_kg_hierarchy,
     get_chat_agent,
+    get_connection_string,
     get_electrophys_feature_tool,
+    get_engine,
     get_httpx_client,
     get_kg_morpho_feature_tool,
+    get_kg_token,
     get_language_model,
     get_literature_tool,
     get_me_model_tool,
     get_morpho_tool,
     get_morphology_feature_tool,
+    get_session,
+    get_settings,
     get_traces_tool,
     get_update_kg_hierarchy,
     get_user_id,
+    get_vlab_and_project,
+    validate_project,
 )
+from neuroagent.app.routers.database.schemas import Base, Threads
 from neuroagent.tools import (
     ElectrophysFeatureTool,
     GetMEModelTool,
@@ -41,6 +54,12 @@ from neuroagent.tools import (
     LiteratureSearchTool,
     MorphologyFeatureTool,
 )
+
+
+def test_get_settings(patch_required_env):
+    settings = get_settings()
+    assert settings.tools.literature.url == "https://fake_url"
+    assert settings.knowledge_graph.url == "https://fake_url/api/nexus/v1/search/query/"
 
 
 @pytest.mark.asyncio
@@ -181,13 +200,144 @@ def test_language_model(monkeypatch, patch_required_env):
     assert language_model.max_tokens == 99
 
 
-def test_get_agent(monkeypatch, patch_required_env):
+@pytest.mark.asyncio
+@pytest.mark.httpx_mock(can_send_already_matched_responses=True)
+async def test_get_vlab_and_project(
+    patch_required_env, httpx_mock, db_connection, monkeypatch
+):
+    # Setup DB with one thread to do the tests
+    monkeypatch.setenv("NEUROAGENT_KEYCLOAK__VALIDATE_TOKEN", "true")
+    test_settings = Settings(
+        db={"prefix": db_connection},
+    )
+    engine = get_engine(test_settings, db_connection)
+    session = next(get_session(engine))
+    user_id = "Super_user"
+    token = "fake_token"
+    httpx_client = AsyncClient()
+    httpx_mock.add_response(
+        url=f"{test_settings.virtual_lab.get_project_url}/test_vlab/projects/test_project",
+        json="test_project_ID",
+    )
+    httpx_mock.add_response(
+        url=f"{test_settings.virtual_lab.get_project_url}/test_vlab_DB/projects/project_id_DB",
+        json="test_project_ID",
+    )
+
+    # create test thread
+    Base.metadata.create_all(bind=engine)
+    new_thread = Threads(
+        user_sub=user_id,
+        vlab_id="test_vlab_DB",
+        project_id="project_id_DB",
+        title="test_title",
+    )
+    session.add(new_thread)
+    session.commit()
+    session.refresh(new_thread)
+
+    try:
+        # Test with info in headers.
+        good_request_headers = Request(
+            scope={
+                "type": "http",
+                "method": "Get",
+                "url": "http://fake_url/thread_id",
+                "headers": [
+                    (b"x-virtual-lab-id", b"test_vlab"),
+                    (b"x-project-id", b"test_project"),
+                ],
+            },
+        )
+        ids = await get_vlab_and_project(
+            user_id=user_id,
+            session=session,
+            request=good_request_headers,
+            settings=test_settings,
+            token=token,
+            httpx_client=httpx_client,
+        )
+        assert ids == {"vlab_id": "test_vlab", "project_id": "test_project"}
+
+        # Test with no infos in headers.
+        bad_request = Request(
+            scope={
+                "type": "http",
+                "method": "GET",
+                "scheme": "http",
+                "server": ("example.com", 80),
+                "path_params": {"dummy_patram": "fake_thread_id"},
+                "headers": [
+                    (b"wong_header", b"wrong value"),
+                ],
+            }
+        )
+        with pytest.raises(HTTPException) as error:
+            await get_vlab_and_project(
+                user_id=user_id,
+                session=session,
+                request=bad_request,
+                settings=test_settings,
+                token=token,
+                httpx_client=httpx_client,
+            )
+        assert (
+            error.value.detail == "thread not found when trying to validate project ID."
+        )
+
+        # Test with no infos in headers, but valid thread_ID.
+        good_request_DB = Request(
+            scope={
+                "type": "http",
+                "method": "GET",
+                "scheme": "http",
+                "server": ("example.com", 80),
+                "path_params": {"thread_id": new_thread.thread_id},
+                "headers": [
+                    (b"wong_header", b"wrong value"),
+                ],
+            }
+        )
+        ids_from_DB = await get_vlab_and_project(
+            user_id=user_id,
+            session=session,
+            request=good_request_DB,
+            settings=test_settings,
+            token=token,
+            httpx_client=httpx_client,
+        )
+        assert ids_from_DB == {"vlab_id": "test_vlab_DB", "project_id": "project_id_DB"}
+
+    finally:
+        # don't forget to close the session, otherwise the tests hangs.
+        session.close()
+
+
+@pytest.mark.asyncio
+async def test_get_agent(monkeypatch, httpx_mock, patch_required_env):
     monkeypatch.setenv("NEUROAGENT_AGENT__MODEL", "simple")
+    monkeypatch.setenv("NEUROAGENT_KEYCLOAK__VALIDATE_TOKEN", "true")
     token = "fake_token"
     httpx_client = AsyncClient()
     settings = Settings()
 
+    vlab_and_project = {"vlab_id": "test_vlab", "project_id": "test_project"}
+    httpx_mock.add_response(
+        url=f'{settings.virtual_lab.get_project_url}/{vlab_and_project["vlab_id"]}/projects/{vlab_and_project["project_id"]}',
+        json="test_project_ID",
+    )
+    valid_project = await validate_project(
+        httpx_client=httpx_client,
+        vlab_id=vlab_and_project["vlab_id"],
+        project_id=vlab_and_project["project_id"],
+        token=token,
+        vlab_project_url=settings.virtual_lab.get_project_url,
+    )
+
     language_model = get_language_model(settings)
+    bluenaas_tool = get_bluenaas_tool(
+        settings=settings, token=token, httpx_client=httpx_client
+    )
     literature_tool = get_literature_tool(
         token=token, settings=settings, httpx_client=httpx_client
     )
@@ -216,7 +366,9 @@ def test_get_agent(monkeypatch, patch_required_env):
     )
 
     agent = get_agent(
+        valid_project,
         llm=language_model,
+        bluenaas_tool=bluenaas_tool,
         literature_tool=literature_tool,
         br_resolver_tool=br_resolver_tool,
         morpho_tool=morpho_tool,
@@ -232,14 +384,33 @@ def test_get_agent(monkeypatch, patch_required_env):
 
 
 @pytest.mark.asyncio
-async def test_get_chat_agent(monkeypatch, db_connection, patch_required_env):
+async def test_get_chat_agent(
+    monkeypatch, db_connection, httpx_mock, patch_required_env
+):
     monkeypatch.setenv("NEUROAGENT_DB__PREFIX", "sqlite://")
+    monkeypatch.setenv("NEUROAGENT_KEYCLOAK__VALIDATE_TOKEN", "true")
 
     token = "fake_token"
     httpx_client = AsyncClient()
     settings = Settings()
 
+    vlab_and_project = {"vlab_id": "test_vlab", "project_id": "test_project"}
+    httpx_mock.add_response(
+        url=f'{settings.virtual_lab.get_project_url}/{vlab_and_project["vlab_id"]}/projects/{vlab_and_project["project_id"]}',
+        json="test_project_ID",
+    )
+    valid_project = await validate_project(
+        httpx_client=httpx_client,
+        vlab_id=vlab_and_project["vlab_id"],
+        project_id=vlab_and_project["project_id"],
+        token=token,
+        vlab_project_url=settings.virtual_lab.get_project_url,
+    )
+
     language_model = get_language_model(settings)
+    bluenaas_tool = get_bluenaas_tool(
+        settings=settings, token=token, httpx_client=httpx_client
+    )
     literature_tool = get_literature_tool(
         token=token, settings=settings, httpx_client=httpx_client
     )
@@ -267,7 +438,9 @@ async def test_get_chat_agent(monkeypatch, db_connection, patch_required_env):
     memory = await anext(get_agent_memory(db_connection))
 
     agent = get_chat_agent(
+        valid_project,
         llm=language_model,
+        bluenaas_tool=bluenaas_tool,
         literature_tool=literature_tool,
         br_resolver_tool=br_resolver_tool,
         morpho_tool=morpho_tool,
@@ -369,3 +542,75 @@ async def test_get_cell_types_kg_hierarchy(
     )
 
     assert os.path.exists(settings.knowledge_graph.ct_saving_path)
+
+
+def test_get_connection_string_full(monkeypatch, patch_required_env):
+    monkeypatch.setenv("NEUROAGENT_DB__PREFIX", "http://")
+    monkeypatch.setenv("NEUROAGENT_DB__USER", "John")
+    monkeypatch.setenv("NEUROAGENT_DB__PASSWORD", "Doe")
+    monkeypatch.setenv("NEUROAGENT_DB__HOST", "localhost")
+    monkeypatch.setenv("NEUROAGENT_DB__PORT", "5000")
+    monkeypatch.setenv("NEUROAGENT_DB__NAME", "test")
+
+    settings = Settings()
+    result = get_connection_string(settings)
+    assert (
+        result == "http://John:Doe@localhost:5000/test"
+    ), "must return fully formed connection string"
+
+
+def test_get_connection_string_no_prefix(monkeypatch, patch_required_env):
+    monkeypatch.setenv("NEUROAGENT_DB__PREFIX", "")
+
+    settings = Settings()
+
+    result = get_connection_string(settings)
+    assert result is None, "should return None when prefix is not set"
+
+
+@patch("neuroagent.app.dependencies.create_engine")
+def test_get_engine(create_engine_mock, monkeypatch, patch_required_env):
+    create_engine_mock.return_value = Mock()
+
+    monkeypatch.setenv("NEUROAGENT_DB__PREFIX", "prefix")
+
+    settings = Settings()
+
+    connection_string = "https://localhost"
+    retval = get_engine(settings=settings, connection_string=connection_string)
+    assert retval is not None
+
+
+@patch("neuroagent.app.dependencies.create_engine")
+def test_get_engine_no_connection_string(
+    create_engine_mock, monkeypatch, patch_required_env
+):
+    create_engine_mock.return_value = Mock()
+
+    monkeypatch.setenv("NEUROAGENT_DB__PREFIX", "prefix")
+
+    settings = Settings()
+
+    retval = get_engine(settings=settings, connection_string=None)
+    assert retval is None
+
+
+@patch("sqlalchemy.orm.Session")
+def test_get_session_success(_):
+    database_url = "sqlite:///:memory:"
+    engine = create_engine(database_url)
+    result = next(get_session(engine))
+    assert isinstance(result, Session)
+
+
+def test_get_session_no_engine():
+    with pytest.raises(HTTPException):
+        next(get_session(None))
+
+
+def test_get_kg_token_with_token(patch_required_env):
+    settings = Settings()
+
+    token = "Test_Token"
+    result = get_kg_token(settings, token)
+    assert result == "Test_Token"