+++ CHANGELOG.md
@@ -7,8 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Switched from OAUTH2 security on FASTAPI to HTTPBearer.
+
 ### Added
 - Add get morphoelectric (me) model tool
+- BlueNaaS simulation tool.
+- Validation of the project ID.
+- BlueNaaS tool test.
 
 ## [0.1.1] - 26.09.2024
 
@@ -20,7 +26,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Update readme
-- Extra multi agent unit tests 
+- Extra multi agent unit tests
+- Extra unit tests for dependencies.py
 
 ### Removed
 - Github action to create the docs.